create searchbar with local search state
